Update IANA timezone DB to 2016d in M51
Cherry pick the following CL to M51 branch

Trunk CL: https://codereview.chromium.org/1985243002 .

What's new in 2016d is found at

  http://mm.icann.org/pipermail/tz-announce/2016-April/000038.html

Rebuilt ICU data/assembly files are checked in (not shown in the
codereview due to their sizes).

While I'm at it, add scripts/LICENSE file that is identical to
LICENSE at the top of the Chromium tree. Because LICENSE in
third_party/icu is for ICU and is not applicable to files in scripts/.

BUG=473288
TBR=mark
TEST=In JavaScript console, run the following.
      apr30_2016_1200 = new Date("04/30/2016 12:00Z")
      may01_2016_1200 = new Date("05/01/2016 12:00Z")
      apr30_2016_1200.toLocaleString("en", {timeZone: "America/Caracas"})
      may01_2016_1200.toLocaleString("en", {timeZone: "America/Caracas"})

  On April 30, 2016, Caracas is 4:30 behind UTC. On May 1, it's 4:00 behind.

Review URL: https://codereview.chromium.org/2016423002 .