Check in the pristine copy of ICU 64rc2
scripts/update.sh 64-rc2 was run to check in the pristine
copy of ICU 64-rc2. When 64.1 is released, the tree will
be updated to that.

Files listed in scripts/data_files_to_preserve were preserved.
Some of them will be deleted (e.g. *local.mk) later because
they're not used any more. To help the transition to a new
data build scheme, they're kept as references.

BUILD.gn and *gyp* files were automatically updated as well.

Bug=944669